refactor: centralize decode-all tile scanning
