+++ CodeGlyphX/AztecCode.Decode.cs
@@ -644,138 +644,13 @@ public static partial class AztecCode {
     }
 
     private static bool TryDecodeAllImageCore(byte[] image, ImageDecodeOptions? options, CancellationToken cancellationToken, out string[] texts) {
-        texts = Array.Empty<string>();
-        if (image is null) throw new ArgumentNullException(nameof(image));
-        var token = ImageDecodeHelper.ApplyBudget(cancellationToken, options, out var budgetCts, out var budgetScope);
-        try {
-            if (token.IsCancellationRequested) return false;
-            if (!ImageReader.TryDecodeRgba32(image, options, out var rgba, out var width, out var height)) return false;
-            var original = rgba;
-            var originalWidth = width;
-            var originalHeight = height;
-            if (!ImageDecodeHelper.TryDownscale(ref rgba, ref width, ref height, options, token)) return false;
-
-            var list = new System.Collections.Generic.List<string>(4);
-            var seen = new System.Collections.Generic.HashSet<string>(StringComparer.Ordinal);
-            CollectAllFromRgba(rgba, width, height, width * 4, token, list, seen);
-            if (!ReferenceEquals(rgba, original) && !token.IsCancellationRequested) {
-                CollectAllFromRgba(original, originalWidth, originalHeight, originalWidth * 4, token, list, seen);
-            }
-
-            if (list.Count == 0) return false;
-            texts = list.ToArray();
-            return true;
-        } finally {
-            budgetCts?.Dispose();
-            budgetScope?.Dispose();
-        }
-    }
-
-    private static void CollectAllFromRgba(byte[] rgba, int width, int height, int stride, CancellationToken token, System.Collections.Generic.List<string> list, System.Collections.Generic.HashSet<string> seen) {
-        if (token.IsCancellationRequested) return;
-        if (AztecDecoder.TryDecode(rgba, width, height, stride, PixelFormat.Rgba32, token, out var text)) {
-            AddUnique(list, seen, text);
-        }
-        ScanTiles(rgba, width, height, stride, token, (tile, tw, th, tstride) => {
-            if (AztecDecoder.TryDecode(tile, tw, th, tstride, PixelFormat.Rgba32, token, out var value)) {
-                AddUnique(list, seen, value);
-            }
-        });
-    }
-
-    private static void AddUnique(System.Collections.Generic.List<string> list, System.Collections.Generic.HashSet<string> seen, string text) {
-        if (string.IsNullOrEmpty(text)) return;
-        if (seen.Add(text)) list.Add(text);
-    }
-
-    private static void ScanTiles(byte[] rgba, int width, int height, int stride, CancellationToken token, Action<byte[], int, int, int> onTile) {
-        if (!CanScanTiles(width, height, stride)) return;
-        var grid = Math.Max(width, height) >= 720 ? 3 : 2;
-        var pad = Math.Max(8, Math.Min(width, height) / 40);
-        var tileW = width / grid;
-        var tileH = height / grid;
-
-        var ctx = new TileScanContext(rgba, width, height, stride, token, onTile, grid, pad, tileW, tileH);
-
-        for (var ty = 0; ty < grid; ty++) {
-            for (var tx = 0; tx < grid; tx++) {
-                if (!ProcessTile(ctx, tx, ty)) return;
-            }
-        }
-    }
-
-    private static bool CanScanTiles(int width, int height, int stride) {
-        return width > 0 && height > 0 && stride >= width * 4;
-    }
-
-    private static bool ProcessTile(TileScanContext ctx, int tx, int ty) {
-        if (ctx.Token.IsCancellationRequested) return false;
-        if (!TryGetTileBounds(ctx, tx, ty, out var x0, out var y0, out var tw, out var th)) return true;
-        if (!TryExtractTile(ctx, x0, y0, tw, th, out var tile, out var tileStride)) return false;
-        ctx.OnTile(tile, tw, th, tileStride);
-        return true;
-    }
-
-    private static bool TryGetTileBounds(TileScanContext ctx, int tx, int ty, out int x0, out int y0, out int tw, out int th) {
-        x0 = tx * ctx.TileW;
-        y0 = ty * ctx.TileH;
-        var x1 = (tx == ctx.Grid - 1) ? ctx.Width : (tx + 1) * ctx.TileW;
-        var y1 = (ty == ctx.Grid - 1) ? ctx.Height : (ty + 1) * ctx.TileH;
-
-        x0 = Math.Max(0, x0 - ctx.Pad);
-        y0 = Math.Max(0, y0 - ctx.Pad);
-        x1 = Math.Min(ctx.Width, x1 + ctx.Pad);
-        y1 = Math.Min(ctx.Height, y1 + ctx.Pad);
-
-        tw = x1 - x0;
-        th = y1 - y0;
-        return tw >= 48 && th >= 48;
-    }
-
-    private static bool TryExtractTile(TileScanContext ctx, int x0, int y0, int tw, int th, out byte[] tile, out int tileStride) {
-        tileStride = tw * 4;
-        tile = new byte[tileStride * th];
-        for (var y = 0; y < th; y++) {
-            if (ctx.Token.IsCancellationRequested) return false;
-            Buffer.BlockCopy(ctx.Rgba, (y0 + y) * ctx.Stride + x0 * 4, tile, y * tileStride, tileStride);
-        }
-        return true;
-    }
-
-    private readonly struct TileScanContext {
-        public byte[] Rgba { get; }
-        public int Width { get; }
-        public int Height { get; }
-        public int Stride { get; }
-        public CancellationToken Token { get; }
-        public Action<byte[], int, int, int> OnTile { get; }
-        public int Grid { get; }
-        public int Pad { get; }
-        public int TileW { get; }
-        public int TileH { get; }
-
-        public TileScanContext(
-            byte[] rgba,
-            int width,
-            int height,
-            int stride,
-            CancellationToken token,
-            Action<byte[], int, int, int> onTile,
-            int grid,
-            int pad,
-            int tileW,
-            int tileH) {
-            Rgba = rgba;
-            Width = width;
-            Height = height;
-            Stride = stride;
-            Token = token;
-            OnTile = onTile;
-            Grid = grid;
-            Pad = pad;
-            TileW = tileW;
-            TileH = tileH;
-        }
+        return DecodeResultHelpers.TryDecodeAllImage(
+            image,
+            options,
+            cancellationToken,
+            (byte[] rgba, int width, int height, int stride, CancellationToken token, out string decoded)
+                => AztecDecoder.TryDecode(rgba, width, height, stride, PixelFormat.Rgba32, token, out decoded),
+            out texts);
     }
 
     /// <summary>
