+++ CodeGlyphX/DataMatrixCode.Decode.cs
@@ -551,84 +551,13 @@ public static partial class DataMatrixCode {
     }
 
     private static bool TryDecodeAllImageCore(byte[] image, ImageDecodeOptions? options, CancellationToken cancellationToken, out string[] texts) {
-        texts = Array.Empty<string>();
-        if (image is null) throw new ArgumentNullException(nameof(image));
-        var token = ImageDecodeHelper.ApplyBudget(cancellationToken, options, out var budgetCts, out var budgetScope);
-        try {
-            if (token.IsCancellationRequested) return false;
-            if (!ImageReader.TryDecodeRgba32(image, options, out var rgba, out var width, out var height)) return false;
-            var original = rgba;
-            var originalWidth = width;
-            var originalHeight = height;
-            if (!ImageDecodeHelper.TryDownscale(ref rgba, ref width, ref height, options, token)) return false;
-
-            var list = new System.Collections.Generic.List<string>(4);
-            var seen = new System.Collections.Generic.HashSet<string>(StringComparer.Ordinal);
-            CollectAllFromRgba(rgba, width, height, width * 4, token, list, seen);
-            if (!ReferenceEquals(rgba, original) && !token.IsCancellationRequested) {
-                CollectAllFromRgba(original, originalWidth, originalHeight, originalWidth * 4, token, list, seen);
-            }
-
-            if (list.Count == 0) return false;
-            texts = list.ToArray();
-            return true;
-        } finally {
-            budgetCts?.Dispose();
-            budgetScope?.Dispose();
-        }
-    }
-
-    private static void CollectAllFromRgba(byte[] rgba, int width, int height, int stride, CancellationToken token, System.Collections.Generic.List<string> list, System.Collections.Generic.HashSet<string> seen) {
-        if (token.IsCancellationRequested) return;
-        if (DataMatrixDecoder.TryDecode(rgba, width, height, stride, PixelFormat.Rgba32, token, out var text)) {
-            AddUnique(list, seen, text);
-        }
-        ScanTiles(rgba, width, height, stride, token, (tile, tw, th, tstride) => {
-            if (DataMatrixDecoder.TryDecode(tile, tw, th, tstride, PixelFormat.Rgba32, token, out var value)) {
-                AddUnique(list, seen, value);
-            }
-        });
-    }
-
-    private static void AddUnique(System.Collections.Generic.List<string> list, System.Collections.Generic.HashSet<string> seen, string text) {
-        if (string.IsNullOrEmpty(text)) return;
-        if (seen.Add(text)) list.Add(text);
-    }
-
-    private static void ScanTiles(byte[] rgba, int width, int height, int stride, CancellationToken token, Action<byte[], int, int, int> onTile) {
-        if (width <= 0 || height <= 0 || stride < width * 4) return;
-        var grid = Math.Max(width, height) >= 720 ? 3 : 2;
-        var pad = Math.Max(8, Math.Min(width, height) / 40);
-        var tileW = width / grid;
-        var tileH = height / grid;
-
-        for (var ty = 0; ty < grid; ty++) {
-            for (var tx = 0; tx < grid; tx++) {
-                if (token.IsCancellationRequested) return;
-                var x0 = tx * tileW;
-                var y0 = ty * tileH;
-                var x1 = (tx == grid - 1) ? width : (tx + 1) * tileW;
-                var y1 = (ty == grid - 1) ? height : (ty + 1) * tileH;
-
-                x0 = Math.Max(0, x0 - pad);
-                y0 = Math.Max(0, y0 - pad);
-                x1 = Math.Min(width, x1 + pad);
-                y1 = Math.Min(height, y1 + pad);
-
-                var tw = x1 - x0;
-                var th = y1 - y0;
-                if (tw < 48 || th < 48) continue;
-
-                var tileStride = tw * 4;
-                var tile = new byte[tileStride * th];
-                for (var y = 0; y < th; y++) {
-                    if (token.IsCancellationRequested) return;
-                    Buffer.BlockCopy(rgba, (y0 + y) * stride + x0 * 4, tile, y * tileStride, tileStride);
-                }
-
-                onTile(tile, tw, th, tileStride);
-            }
-        }
+        return DecodeResultHelpers.TryDecodeAllImage(
+            image,
+            options,
+            cancellationToken,
+            (byte[] rgba, int width, int height, int stride, CancellationToken token, out string decoded)
+                => DataMatrixDecoder.TryDecode(rgba, width, height, stride, PixelFormat.Rgba32, token, out decoded),
+            out texts);
     }
 
     /// <summary>
